Apparent variable overflow
Checking out my Arduino Mini I found that sketch wasn't running correctly. For example, it did Setup(), Loop()1, Loop()2 and then Setup() again, so, it was doing an unexpected Reset. By the use of flags (multiple Serial.print) I pointed the error to sendall() code. Then, printing allData variable, I found an error on the char array during the second loop, so I thought it was like a memory overflow an tried with free() immediately after malloc() and worked very fine. I don't know why the free() at the end of the process is not working. Thanks for your work!